[#370] Activation on PHP < 5.6 now displays nice error message
